requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloud-sql-python-connector==0.8.1
cryptography==38.0.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.1
google-auth==2.12.0
greenlet==1.1.3
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
pg8000==1.29.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
scramp==1.4.1
six==1.16.0
SQLAlchemy==1.4.41
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
